AB#80 Extract runScript.ts
